Add repartition test for chained aggregates to ensure consistent hash distribution
