+++ datafusion/core/tests/physical_optimizer/enforce_distribution.rs
@@ -3834,3 +3834,62 @@ fn test_replace_order_preserving_variants_with_fetch() -> Result<()> {
 
     Ok(())
 }
+
+#[tokio::test]
+async fn repartition_between_chained_aggregates() -> Result<()> {
+    // Build a two-partition, empty MemTable with the expected schema to mimic the
+    // reported failing plan: Sort -> Aggregate(ts, region) -> Sort -> Aggregate(ts).
+    use arrow::datatypes::{DataType, Field, Schema};
+    use arrow::record_batch::RecordBatch;
+    use datafusion_expr::col;
+    use datafusion_functions_aggregate::expr_fn::count;
+    use datafusion_physical_plan::{collect, displayable};
+
+    let schema = Arc::new(Schema::new(vec![
+        Field::new("ts", DataType::Int64, true),
+        Field::new("region", DataType::Utf8, true),
+        Field::new("value", DataType::Int64, true),
+    ]));
+    let empty_batch = RecordBatch::new_empty(schema.clone());
+    let partitions = vec![vec![empty_batch.clone()], vec![empty_batch]];
+    let mem_table = MemTable::try_new(schema, partitions)?;
+
+    let config = SessionConfig::new().with_target_partitions(2);
+    let ctx = SessionContext::new_with_config(config);
+    ctx.register_table("metrics", Arc::new(mem_table))?;
+
+    let df = ctx
+        .table("metrics")
+        .await?
+        .sort(vec![
+            col("ts").sort(true, true),
+            col("region").sort(true, true),
+        ])?
+        .aggregate(vec![col("ts"), col("region")], vec![count(col("value"))])?
+        .sort(vec![
+            col("ts").sort(true, true),
+            col("region").sort(true, true),
+        ])?
+        .aggregate(vec![col("ts")], vec![count(col("region"))])?;
+
+    let physical_plan = df.create_physical_plan().await?;
+
+    // The optimizer should either keep the stream single-partitioned via the
+    // sort-preserving merge, or insert a repartition between the two aggregates
+    // so that the second grouping sees a consistent hash distribution. Either
+    // path protects against the panic that was previously reported for this
+    // plan shape.
+    let plan_display = displayable(physical_plan.as_ref()).indent(true).to_string();
+    let has_repartition =
+        plan_display.contains("RepartitionExec: partitioning=Hash([ts@0], 2)");
+    assert!(
+        has_repartition || plan_display.contains("SortPreservingMergeExec"),
+        "Expected either a repartition between aggregates or a sort-preserving merge chain"
+    );
+
+    // Execute the optimized plan to ensure the empty, multi-partition pipeline does not panic.
+    let batches = collect(physical_plan, ctx.task_ctx()).await?;
+    assert!(batches.is_empty());
+
+    Ok(())
+}
